+++ src/ethereum/info/arbitrumInfo.ts
@@ -151,6 +151,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.etherscan.io', 'https://api.arbiscan.io']
     },
     {