+++ src/ethereum/info/avalancheInfo.ts
@@ -197,6 +197,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: [
         'https://api.etherscan.io',
         'https://api.avascan.info/v2/network/mainnet/evm/43114/etherscan',