+++ src/ethereum/info/baseInfo.ts
@@ -69,6 +69,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.etherscan.io', 'https://api.basescan.org']
     },
     {