+++ src/ethereum/info/ethereumInfo.ts
@@ -1168,6 +1168,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.etherscan.io']
     },
     {