+++ src/ethereum/info/hyperEvmInfo.ts
@@ -131,6 +131,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.routescan.io/v2/network/mainnet/evm/999/etherscan']
     }
   ],