+++ src/ethereum/info/optimismInfo.ts
@@ -189,6 +189,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.etherscan.io']
     }
   ],