+++ src/ethereum/info/sonicInfo.ts
@@ -144,6 +144,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: ['https://api.etherscan.io', 'https://api.sonicscan.org']
     }
   ],