+++ src/ethereum/info/zksyncInfo.ts
@@ -96,6 +96,7 @@ const networkInfo: EthereumNetworkInfo = {
     },
     {
       type: 'evmscan',
+      gastrackerSupport: true,
       servers: [
         'https://api.etherscan.io',
         'https://block-explorer-api.mainnet.zksync.io',