+++ src/ethereum/networkAdapters/EvmScanAdapter.ts
@@ -59,6 +59,9 @@ const NUM_TRANSACTIONS_TO_QUERY = 50
 export interface EvmScanAdapterConfig {
   type: 'evmscan'
   servers: string[]
+
+  /** Whether the API supports gastracker module */
+  gastrackerSupport?: boolean
 }
 
 export class EvmScanAdapter extends NetworkAdapter<EvmScanAdapterConfig> {
@@ -308,7 +311,12 @@ export class EvmScanAdapter extends NetworkAdapter<EvmScanAdapterConfig> {
     }
 
     const response = await this.ethEngine.fetchCors(`${url}${apiKeyParam}`)
-    if (!response.ok) this.throwError(response, 'fetchGetEtherscan', url)
+
+    if (!response.ok) {
+      const resBody = await response.text()
+      this.throwError(response, 'fetchGetEtherscan', url, resBody)
+    }
+
     const data = await response.json()
     const cleanData = asEvmScanResponse(asUnknown)(data)
     if (