+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- fixed: (EVM) Fix missing rate-limit error detection
+- added: (EVM) Add gastrackerSupport flag to evmscan adapter config
+- added: (EVM) Include more cases to RateLimitError
+
 ## 4.55.5 (2025-07-28)
 
 - added: Implement ethFeeHistory fee algorithm for EIP-1559 chains with percentile-based priority fee estimation, initially enabled for Botanix network